@@ -29,38 +29,38 @@
   },
   "dependencies": {
     "@ledgerhq/hw-transport-node-hid": "6.29.8",
-    "@zondax/ledger-filecoin": "^2.1.1-beta",
-    "@zondax/zemu": "^0.55.3"
+    "@zondax/ledger-filecoin": "^3.0.0",
+    "@zondax/zemu": "^0.55.3",
+    "varint": "^6.0.0"
   },
   "devDependencies": {
     "@ethereumjs/common": "^10.0.0",
     "@ethereumjs/tx": "^10.0.0",
-    "@ledgerhq/hw-app-eth": "6.42.4",
+    "@ledgerhq/hw-app-eth": "6.45.10",
     "@types/elliptic": "^6.4.18",
     "@types/jest": "^30.0.0",
-    "@types/ledgerhq__hw-transport": "^4.21.8",
-    "@types/node": "^24.0.4",
+    "@types/node": "^24.0.14",
     "@types/secp256k1": "^4.0.6",
-    "@typescript-eslint/eslint-plugin": "^8.35.1",
-    "@typescript-eslint/parser": "^8.35.1",
+    "@typescript-eslint/eslint-plugin": "^8.37.0",
+    "@typescript-eslint/parser": "^8.37.0",
     "blakejs": "^1.2.1",
     "bn.js": "^5.2.2",
     "crypto-js": "4.2.0",
     "elliptic": "^6.6.1",
-    "eslint": "^9.29.0",
+    "eslint": "^9.31.0",
     "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-jest": "^28.8.0",
-    "eslint-plugin-prettier": "^5.5.0",
+    "eslint-plugin-prettier": "^5.5.1",
     "eslint-plugin-tsdoc": "^0.4.0",
     "eslint-plugin-unused-imports": "^4.1.4",
-    "jest": "30.0.3",
-    "jest-runner": "^30.0.3",
+    "jest": "30.0.4",
+    "jest-runner": "^30.0.4",
     "jest-serial-runner": "^1.2.2",
     "js-sha256": "0.11.1",
     "prettier": "^3.6.2",
     "secp256k1": "^5.0.1",
-    "sort-package-json": "^3.3.0",
+    "sort-package-json": "^3.4.0",
     "ts-jest": "^29.4.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
